fixed burn tx test case for timelock and added more test cases

+++ src/bridge/components/helper.rs
@@ -11,7 +11,7 @@ pub fn generate_burn_script_address() -> Address {
     Address::p2wsh(&generate_burn_script(), Network::Testnet)
 }
 
-pub fn generate_timelock_script(pubkey: &XOnlyPublicKey, weeks: i64) -> Script {
+pub fn generate_timelock_script(pubkey: &XOnlyPublicKey, weeks: u32) -> Script {
     script! {
       { NUM_BLOCKS_PER_WEEK * weeks }
       OP_CSV
@@ -21,7 +21,7 @@ pub fn generate_timelock_script(pubkey: &XOnlyPublicKey, weeks: i64) -> Script {
     }
 }
 
-pub fn generate_timelock_script_address(pubkey: &XOnlyPublicKey, weeks: i64) -> Address {
+pub fn generate_timelock_script_address(pubkey: &XOnlyPublicKey, weeks: u32) -> Address {
     Address::p2wsh(&generate_timelock_script(pubkey, weeks), Network::Testnet)
 }
 
@@ -38,4 +38,4 @@ pub fn generate_pay_to_pubkey_script_address(pubkey: &XOnlyPublicKey) -> Address
 
 pub type Input = (OutPoint, Amount);
 
-pub const NUM_BLOCKS_PER_WEEK: i64 = 1008;
+pub const NUM_BLOCKS_PER_WEEK: u32 = 1008;
